Read Post Data as Byte Array
Post Data was being read as Http Headers, when it should be returned as
a byte array. Applications may post different data in different formats,
and the library should allow the host application to decide how to
decode.

I may rewrite to handle post data as a stream for large posts, as well
as add a limiter for large posts/security/malformed posts at a later
point in time.